[A5] Started implementing client
- Implemented data types and Get(), Put(), Append() functions.
- Skeleton for handleOp(), but not yet implemented.

+++ A5-key-value-store/src/kvraft/common.go
@@ -19,7 +19,7 @@ const (
 
 // Unique IDs for Ops. 2-tuple (Client ID, Client Op ID)
 type OpId struct {
-	ClientId   int
+	ClientId   int64
 	ClientOpId int
 }
